modified oracle-cep (removed from maven) and added common-domain.jar to ocep
We need the common domain classes in oracle cep. But building this with
maven won’t work. at least i dont know how for now and no time to
figure it out. so we just take the built common-domain and add it as a
jar dependency to the oracle cep.